% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbdf.R
\name{new_vbdf}
\alias{new_vbdf}
\title{Constructor for class vbdf}
\usage{
new_vbdf(x, bloc_var = character(), var_type = c("discrete", "continuous"))
}
\arguments{
\item{x}{a data.frame}

\item{bloc_var}{character vector naming the variables to define voting blocs}

\item{var_type}{string, the type, discrete or continuous}
}
\description{
Constructor for class vbdf
}
