% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_vbdiff}
\alias{plot_vbdiff}
\alias{plot.vbdiff}
\title{Plot voting blocs' year-to-year difference}
\usage{
plot_vbdiff(
  vbdiff,
  x = get_bloc_var(vbdiff),
  y,
  ymin,
  ymax,
  discrete = length(unique(vbdiff[[x]])) < 20
)

\method{plot}{vbdiff}(
  vbdiff,
  x = get_bloc_var(vbdiff),
  y,
  ymin,
  ymax,
  discrete = length(unique(vbdiff[[x]])) < 20
)
}
\arguments{
\item{vbdiff}{a data.frame of year-to-year differences in the format of the output from \link{vb_diff}.}

\item{x}{string naming the column to plot on the x axis.}

\item{y}{string naming the column to plot on the y axis.}

\item{ci_low}{string naming the column to plot as the lower bound of the confidence interval.}

\item{ci_high}{string naming the column to plot as the upper bound of the confidence interval.}
}
\description{
Plot voting blocs' year-to-year difference
}
