% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difference.R
\name{vb_difference}
\alias{vb_difference}
\title{Calculate differences in bloc contributions}
\usage{
vb_difference(
  vbdf,
  diff_col = intersect(names(vbdf), c("prob", "pr_turnout", "pr_votedem", "pr_voterep",
    "cond_rep", "net_rep")),
  sort_col = "year"
)
}
\arguments{
\item{vbdf}{data.frame holding the results of voting bloc analyses.}

\item{diff_col}{character vector naming the column(s) in \code{vbdf} with
which to compute differences.}

\item{sort_col}{character vector naming the column(s) in \code{vbdf} to use
for sorting before calling \link[base]{diff}.}
}
\value{
A \code{vbdiff} object, similar to \code{vbdf} plus two columns:
  \code{diff_*}, which holds the difference in \code{vbdf[[diff_col]]} across
  consecutive years in \code{vbdf$year}; and \code{comp}, which holds a
  string tag for the years compared.

A \code{vbdiff} object.
}
\description{
Use vbdf output to calculate differences
in blocs' net Republican vote contributions.
}
