% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difference.R
\name{vb_hypothetical}
\alias{vb_hypothetical}
\title{Calculate hypothetical bloc-contribution differences}
\usage{
vb_hypothetical(x, y, estimates)
}
\arguments{
\item{x, y}{two sets of vbdf objects with matching column names}

\item{diff_col}{character vector naming the column(s) in \code{vbdf} with
which to compute differences.}

\item{sort_col}{character vector naming the column(s) in \code{vbdf} to use
for sorting before calling \link[base]{diff}.}
}
\description{
Calculate hypothetical bloc-contribution differences
}
