% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty.R
\name{vb_uncertainty}
\alias{vb_uncertainty}
\title{Compute uncertainty for voting blocs analysis}
\usage{
vb_uncertainty(object, ...)
}
\arguments{
\item{object}{a \code{vbdf} or \code{vbdiff} object.}

\item{...}{further arguments to pass to methods}
}
\description{
Compute uncertainty for voting blocs analysis
}
