% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty.R
\name{vb_uncertainty.vbdf}
\alias{vb_uncertainty.vbdf}
\title{Compute uncertainty for a vbdf object}
\usage{
\method{vb_uncertainty}{vbdf}(
  vbdf,
  type = c("continuous", "binned", "discrete"),
  estimates,
  na.rm = FALSE,
  funcs = c("original", "mean", "median", "low", "high"),
  low_ci = 0.025,
  high_ci = 0.975,
  bin_col
)
}
\arguments{
\item{vbdf}{a \code{vbdf} object, usually the result of \code{vb_discrete} or
\code{vb_continuous}.}

\item{type}{a string naming the type of independent variable summary. Use
\code{"binned"} when working with binned output of \code{vb_continuous()}.}

\item{estimates}{character vector naming columns for which to calculate
uncertainty estimates.}

\item{na.rm}{logical indicating whether to remove \code{NA} values in
\code{estimates}.}

\item{funcs}{character vector of summary functions to apply to
\code{estimates}. Alternatively, supply your own list of functions, which
should accept a numeric vector input and return a scalar.}

\item{low_ci}{numeric. If you include the string \code{"low"} in \code{funcs}, then use this argument to control the lower bound of the confidence interval.}

\item{high_ci}{numeric. If you include the string \code{"high"} in \code{funcs}, then use this argument to control the upper bound of the confidence interval.}

\item{bin_col}{character vector naming the columns that contain the bins. Only used if  \code{type} is \code{"binned"}.}
}
\value{
a \code{vbdf} object with additional columns for each combination
  of \code{estimates} and \code{funcs}.
}
\description{
Compute uncertainty for a vbdf object
}
