% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty.R
\name{vb_uncertainty.vbdiff}
\alias{vb_uncertainty.vbdiff}
\title{Compute uncertainty for a vbdiff object}
\usage{
\method{vb_uncertainty}{vbdiff}(
  vbdiff,
  type = c("continuous", "binned", "discrete"),
  estimates,
  bin_col,
  na.rm = FALSE,
  funcs = c("original", "mean", "median", "low", "high"),
  low_ci = 0.025,
  high_ci = 0.975
)
}
\arguments{
\item{vbdiff}{a \code{vbdiff} object, the result of \code{vb_difference()}.}

\item{type}{a string naming the type of independent variable summary. Use
\code{"binned"} when working with binned output of \code{vb_continuous()}.}

\item{estimates}{character vector naming columns for which to calculate
uncertainty estimates.}

\item{bin_col}{character vector naming columns that contain the bins. Used
only if  \code{type} is \code{"binned"}.}

\item{na.rm}{logical indicating whether to remove \code{NA} values in
\code{estimates}.}

\item{funcs}{character vector of summary functions to apply to
\code{estimates}. Alternatively, supply your own list of functions, which
should accept a numeric vector input and return a scalar.}

\item{low_ci}{numeric. If you include the string \code{"low"} in
\code{funcs}, then use this argument to control the lower bound of the
confidence interval.}

\item{high_ci}{numeric. If you include the string \code{"high"} in
\code{funcs}, then use this argument to control the upper bound of the
confidence interval.}
}
\value{
a \code{vbdiff} object with additional columns for each combination
  of \code{estimates} and \code{funcs}.
}
\description{
Compute uncertainty for a vbdiff object
}
