% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbdf.R
\name{vbdf}
\alias{vbdf}
\title{Create a vbdf object}
\usage{
vbdf(data, bloc_var, var_type = c("discrete", "continuous"))
}
\arguments{
\item{data}{data.frame of voting-bloc results to convert to a \code{vbdf} object}

\item{bloc_var}{string, the name of the variable that defines the voting blocs}

\item{var_type}{string, the type of variable, discrete or continuous}
}
\value{
A \code{vbdf} object.
}
\description{
Create a vbdf object holding bloc-level estimates of composition, turnout,
and/or vote choice. This function is mostly for internal use, but you may
want it to create a \code{vbdf} object from your own voting bloc analysis.
A valid \code{vbdf} object can be used in [vb_difference] and [vb_plot].
}
