% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbdf.R
\name{check_vbdf}
\alias{check_vbdf}
\title{Validator for class vbdf}
\usage{
check_vbdf(x, tolerance = sqrt(.Machine$double.eps))
}
\arguments{
\item{x}{object to check}

\item{tolerance}{tolerance used when checking range of probability estimates}
}
\description{
Validator for class vbdf
}
