% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous.R
\name{estimate_density}
\alias{estimate_density}
\title{Estimate density}
\usage{
estimate_density(x, min, max, n_points = 100, w = NULL, ...)
}
\arguments{
\item{x}{numeric vector or matrix}

\item{min}{numeric vector giving the lower bound of evaluation points for each variable in \code{x}}

\item{max}{numeric vector giving the upper bound of evaluation points for each variable in \code{x}}

\item{n_points}{number of evaluation points (estimates)}

\item{w}{vector of weights. Default uses uniform weighting.}

\item{...}{further arguments to pass to \link[ks]{kde}}
}
\description{
Run \link[ks]{kde} for weighted density estimation
of a  \code{x} at \code{n_points}
evenly spaced points between \code{min} and {max}.
}
