% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbdiff.R
\name{new_vbdiff}
\alias{new_vbdiff}
\title{vbdiff class and methods
Constructor for class vbdiff}
\usage{
new_vbdiff(x, bloc_var, var_type, diff_col)
}
\description{
vbdiff class and methods
Constructor for class vbdiff
}
